Safety Backup

+++ src/main/java/combit/hu/porphyr/service/DeveloperService.java
@@ -26,30 +26,57 @@ public class DeveloperService {
         return developerRepository.findAll();
     }
 
+    /**
+    * Hibalehetőségek: <br/>
+    * - Nincs kitöltve a név <br/>
+    * - Már van ilyen nevű fejlesztő <br/>
+    */
     public void insertNewDeveloper(final @NonNull DeveloperEntity newDeveloperEntity) {
-        developerRepository.save(newDeveloperEntity);
-    }
-
-    public void modifyDeveloper(final @NonNull DeveloperEntity modifiedDeveloper) {
-        Long developerID = modifiedDeveloper.getId();
-        if( developerID != null) {
-            developerRepository.save( modifiedDeveloper );
+        if (newDeveloperEntity.getName().isEmpty()){
+            throw( new ServiceException( ServiceException.Exceptions.DEVELOPER_WITH_EMPTY_NAME_CANT_INSERT) );
+        } else if ( ! developerRepository.findAllByName(newDeveloperEntity.getName()).isEmpty() ) {
+            throw( new ServiceException( ServiceException.Exceptions.DEVELOPER_WITH_SAME_NAME_CANT_INSERT) );
         } else {
-            throw( new ServiceException( ServiceException.Exceptions.DEVELOPER_NOT_SAVED_CANT_MODIFY) );
+            developerRepository.save(newDeveloperEntity);
         }
     }
 
+    /**
+     * Hibalehetőségek: <br/>
+     * - Nincs kitöltve az ID <br/>
+     * - Nincs kitöltve a név <br/>
+     * - Már van ilyen nevű fejlesztő <br/>
+     */
+    public void modifyDeveloper(final @NonNull DeveloperEntity modifiedDeveloper) {
+        if( modifiedDeveloper.getId() == null) {
+            throw (new ServiceException(ServiceException.Exceptions.DEVELOPER_NOT_SAVED_CANT_MODIFY));
+        } else if( modifiedDeveloper.getName().isEmpty()) {
+            throw (new ServiceException(ServiceException.Exceptions.DEVELOPER_WITH_EMPTY_NAME_CANT_MODIFY));
+        } else if ( ! developerRepository.findAllByName(modifiedDeveloper.getName()).isEmpty() ) {
+            throw( new ServiceException( ServiceException.Exceptions.DEVELOPER_WITH_SAME_NAME_CANT_MODIFY) );
+        } else {
+            developerRepository.save(modifiedDeveloper);
+        }
+    }
+
+    /**
+     * Hibalehetőségek: <br/>
+     * - Nincs kitöltve az ID <br/>
+     * - A fejlesztő hozzá van rendelve egy vagy több projekthez
+     */
     public void deleteDeveloper(final @NonNull DeveloperEntity developerEntity) {
         Long developerID = developerEntity.getId();
+        DeveloperEntity actualDeveloperEntity;
         if( developerID != null) {
-            developerRepository.deleteById( developerID );
+            actualDeveloperEntity = developerRepository.findAllById( developerEntity.getId() );
+            if (actualDeveloperEntity.getDeveloperProjects().isEmpty()) {
+                developerRepository.deleteById(developerID);
+            } else {
+                throw( new ServiceException( ServiceException.Exceptions.DEVELOPER_WITH_PROJECTS_CANT_DELETE) );
+            }
         } else {
             throw( new ServiceException( ServiceException.Exceptions.DEVELOPER_NOT_SAVED_CANT_DELETE) );
         }
-    }
-
-    public void deleteAllDevelopers() {
-        developerRepository.deleteAll();
     }
 
 }
